feat: events for channel opening and closing (#22)
* feat: events for channel opening and closing

The event is generated as soon as a channel opening or closing has been
initiated. Pending channels are evaluated for this purpose.
The history of pending channels is monitored in a separate pending
channel manager. The status of the manager is updated ad hoc when a
pending open channel event or an on-chain event occurs. It is also
updated periodically to ensure that the status is not lost,
particularly in the case of external wallets.

* docs: fix example

* fix: remaks from code review

* New template variables for `channel_close_event`

CloseInitiator and CloseType